+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * New command `report tag` to list tags with time statistics (#93)
 
+### Bugfixes
+
+* Fix #94 broken child project time aggregation (#95)
+
 ## [[v0.3.2]](https://github.com/mlange-42/track/compare/v0.3.1...v0.3.2)
 
 ### Features